+++ .eslintrc
@@ -35,8 +35,6 @@
     "camelcase": 0,
     "class-methods-use-this": 0,
     "consistent-return": 0,
-    // "default-case": 1,
-    "import/no-cycle": 0,
     "jsx-a11y/click-events-have-key-events": 0,
     "jsx-a11y/label-has-associated-control": 0,
     "jsx-a11y/no-noninteractive-element-interactions": 0,
@@ -50,22 +48,14 @@
     "no-continue": 0,
     "no-console": 0,
     "no-mixed-operators": 0,
-    "no-multi-assign": 0,
     "no-nested-ternary": 0,
     "no-param-reassign": 0,
-    "no-plusplus": 0,
     "no-prototype-builtins": 0,
     "no-restricted-properties": 0,
     "no-restricted-syntax": 0,
-    "no-return-assign": 0,
-    "no-return-await": 0,
-    "no-sequences": 0,
     "no-shadow": 0,
-    "no-sparse-arrays": 0,
     "no-underscore-dangle": 0,
-    "no-unused-expressions": 0,
     "no-void": 0,
-    // "react/destructuring-assignment": 1,
     "react/jsx-filename-extension": [
       2,
       {
