+++ src/SyncStatus.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
+import React, { FunctionComponent } from 'react';
 import { IconStatus } from './icons/IconStatus';
-import { ForceSyncronize } from './App';
+import type { ForceSyncronize } from './App';
 import './SyncStatus.css';
 
 interface SyncStatusProps {
@@ -9,7 +9,7 @@ interface SyncStatusProps {
   synchronized: boolean;
 }
 
-export default function SyncStatus(props: SyncStatusProps) {
+export const SyncStatus: FunctionComponent<SyncStatusProps> = (props: SyncStatusProps) => {
   const {
     connected,
     forceSync,
@@ -36,4 +36,4 @@ export default function SyncStatus(props: SyncStatusProps) {
       />
     </div>
   );
-}
+};
