+++ src/lib/roundToNearest5.ts
@@ -1,0 +1,8 @@
+/**
+ * A helper function that rounds to the nearest 5.
+ *
+ * @param number
+ */
+export function roundToNearest5(number: number) {
+  return 5 * Math.round(number / 5);
+}
