+++ src/viewer/MaterialScoresViewer.tsx
@@ -3,7 +3,7 @@ import {
   toDisplayCutSpeed,
   toDisplayPowerWords,
 } from '../lib/glowforgeUnits';
-import { PluginMaterial } from '../material/materialPlugin';
+import type { PluginMaterial } from '../material/materialPlugin';
 import { precisionRound } from '../lib/utils';
 import './MaterialViewer.css';
