+++ src/viewer/QrCodeViewer.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { PluginMaterial } from '../material/materialPlugin';
+import type { PluginMaterial } from '../material/materialPlugin';
 import { sha1 } from '../lib/crypto';
 import { qrcodeAsDataUri, qrcodeAsSvg } from '../lib/qrCode';
-import { toTinyMaterial } from '../material/material';
+import { toTinyMaterial } from '../material/materialConversion';
 import './MaterialViewer.css';
 
 interface QrCodeViewerProps {
